add parseFloat

+++ js/task-3.js
@@ -1,5 +1,5 @@
 function getElementWidth (content, padding, border) {
-    const contentWidth = parseInt (content); 
+    const contentWidth = parseFloat (content); 
     const paddingWidth = parseFloat (padding);
     const borderWidth = parseFloat (border);
     const totalWidth = contentWidth + (paddingWidth + borderWidth)*2;
